% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch-processor.R
\name{batch_processor}
\alias{batch_processor}
\title{Batch Genome Processor}
\usage{
batch_processor(accessions, cache_dir = NULL)
}
\arguments{
\item{accessions}{Vector of genome accession numbers}

\item{cache_dir}{Directory to cache downloaded data}
}
\value{
A batch processor object
}
\description{
Process multiple genomes in batch with progress tracking and error handling.
}
\examples{
\dontrun{
processor <- batch_processor(c("NC_044967.1", "NC_044968.1"))
results <- processor$process_all()
}
}
