% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database-management.R
\name{clean_empty_dirs}
\alias{clean_empty_dirs}
\title{Clean empty directories in database}
\usage{
clean_empty_dirs(db_path = NULL)
}
\arguments{
\item{db_path}{Optional path to genome database. If NULL, uses current database.}
}
\value{
Number of directories removed
}
\description{
Remove empty subdirectories from the genome database
}
\examples{
\dontrun{
clean_empty_dirs()
}
}
