% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome-manager.R
\name{.check_existing_genome}
\alias{.check_existing_genome}
\title{Check if genome already exists in database}
\usage{
.check_existing_genome(accession, db_path, overwrite, interactive)
}
\description{
Check if genome already exists in database
}
\keyword{internal}
