% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome-manager.R
\name{.perform_import}
\alias{.perform_import}
\title{Perform the actual import operation}
\usage{
.perform_import(import_path, db_path, accession, validation)
}
\description{
Perform the actual import operation
}
\keyword{internal}
