% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome-manager.R
\name{.prompt_incomplete_data}
\alias{.prompt_incomplete_data}
\title{Prompt user about incomplete data}
\usage{
.prompt_incomplete_data(validation, accession, db_path)
}
\description{
Prompt user about incomplete data
}
\keyword{internal}
