% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome-manager.R
\name{.validate_export_data}
\alias{.validate_export_data}
\title{Validate exported data completeness}
\usage{
.validate_export_data(import_path, accession)
}
\description{
Validate exported data completeness
}
\keyword{internal}
