% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome-manager.R
\name{export_genome}
\alias{export_genome}
\title{Export genome data for sharing}
\usage{
export_genome(
  accession,
  export_dir = NULL,
  db_path = NULL,
  include_sequence = TRUE,
  include_features = TRUE
)
}
\arguments{
\item{accession}{Genome accession number}

\item{export_dir}{Directory where the export package will be created.
If NULL, creates in current working directory.}

\item{db_path}{Optional path to genome database. If NULL, uses current database.}

\item{include_sequence}{Whether to include sequence data in the export}

\item{include_features}{Whether to include feature data in the export}
}
\value{
Path to the created export directory
}
\description{
Export genome data to a portable format that can be shared with colleagues.
Creates a standardized archive containing all data for the specified genome.
}
\examples{
\dontrun{
# Export all data for a genome
export_genome("ON400500.1", "./exports")

# Export only metadata
export_genome("ON400500.1", "./exports",
              include_sequence = FALSE, include_features = FALSE)
}
}
