% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_download.R
\name{fgenome_download}
\alias{fgenome_download}
\title{Download genome metadata and sequence from NCBI}
\usage{
fgenome_download(genome_id = "ON400500.1")
}
\arguments{
\item{genome_id}{The accession number of the genome (e.g., "ON400500.1")}
}
\value{
A data.frame containing genome metadata and sequence
}
\description{
Download genome metadata and sequence from NCBI
}
\examples{
\dontrun{
Download a single genome
fgenome_download("ON400500.1")
}
}
