% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_download.R
\name{ft_download}
\alias{ft_download}
\title{Download feature table for a genome from NCBI}
\usage{
ft_download(genome_id = "ON400500.1")
}
\arguments{
\item{genome_id}{The accession number of the genome (e.g., "ON400500.1")}
}
\value{
A data.frame containing feature table information
}
\description{
Download feature table for a genome from NCBI
}
\examples{
\dontrun{
ft_download("ON400500.1")
}
}
