% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database-management.R
\name{get_default_genome_db}
\alias{get_default_genome_db}
\title{Get default genome database path}
\usage{
get_default_genome_db()
}
\value{
Path to the default genome database directory
}
\description{
Get default genome database path
}
\keyword{internal}
