% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database-management.R
\name{get_genome_db}
\alias{get_genome_db}
\title{Get current genome database path}
\usage{
get_genome_db()
}
\value{
Path to the current genome database directory
}
\description{
Get current genome database path
}
\examples{
\dontrun{
get_genome_db()
}
}
