% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database-management.R
\name{list_genomes}
\alias{list_genomes}
\title{List available genomes in database}
\usage{
list_genomes(db_path = NULL)
}
\arguments{
\item{db_path}{Optional path to genome database. If NULL, uses current database.}
}
\value{
Data frame with information about available genomes
}
\description{
List available genomes in database
}
\examples{
\dontrun{
list_genomes()
list_genomes("~/my_genomes")
}
}
