% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database-management.R
\name{remove_genome}
\alias{remove_genome}
\title{Remove genome from database}
\usage{
remove_genome(accession, db_path = NULL)
}
\arguments{
\item{accession}{Genome accession number}

\item{db_path}{Optional path to genome database. If NULL, uses current database.}
}
\value{
Logical indicating success
}
\description{
Remove genome from database
}
\examples{
\dontrun{
remove_genome("ON400500.1")
}
}
