% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{search_ncbi_accessions}
\alias{search_ncbi_accessions}
\title{Search NCBI for accessions}
\usage{
search_ncbi_accessions(search_term, api_key = NULL)
}
\description{
Search NCBI for accessions
}
\keyword{internal}
