% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database-management.R
\name{set_genome_db}
\alias{set_genome_db}
\title{Set genome database directory}
\usage{
set_genome_db(path = NULL, create = TRUE)
}
\arguments{
\item{path}{Path to the genome database directory. If NULL, uses the
default location in the R user data directory.}

\item{create}{Logical. Whether to create the directory if it doesn't exist.}
}
\value{
The path to the genome database directory (invisibly)
}
\description{
Set the directory where genome data will be stored. This allows users
to specify a persistent location for their genome databases.
}
\examples{
\dontrun{
# Set to a custom directory
set_genome_db("~/my_genome_database")

# Set to a project-specific directory
set_genome_db("./genome_data")

# Reset to default location
set_genome_db()
}
}
