% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome-manager.R
\name{share_genome}
\alias{share_genome}
\title{Create shareable genome package}
\usage{
share_genome(accession, output_file = NULL, db_path = NULL)
}
\arguments{
\item{accession}{Genome accession number}

\item{output_file}{Path to the output archive file (.zip or .tar.gz)}

\item{db_path}{Optional path to genome database. If NULL, uses current database.}
}
\value{
Path to the created archive file
}
\description{
Create a compressed archive of genome data for easy sharing.
}
\examples{
\dontrun{
# Create a zip archive
share_genome("ON400500.1", "./ON400500.1.zip")
}
}
