% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-genomes.R
\name{update_genomes}
\alias{update_genomes}
\title{Incremental Update Genomes}
\usage{
update_genomes(search_term = NULL, cache_dir = NULL)
}
\arguments{
\item{search_term}{NCBI search term for genomes}

\item{cache_dir}{Directory to cache downloaded data}
}
\value{
An update system object
}
\description{
Check for and download new genomes from NCBI. Default for ASFV
}
\examples{
\dontrun{
update_genomes <- update_genomes('"African swine fever virus"[Organism]')
update_info <- update_genomes$check_updates()
if (update_info$needs_update) {
  update_genomes$perform_update()
}
}
}
